Transformers  Import fixes

+++ simpletransformers/classification/transformer_models/roberta_model.py
@@ -77,7 +77,8 @@ class RobertaForSequenceClassification(BertPreTrainedModel):
                 loss = loss_fct(logits.view(-1), labels.view(-1))
             else:
                 loss_fct = CrossEntropyLoss(weight=self.weight)
-                loss = loss_fct(logits.view(-1, self.num_labels), labels.view(-1))
+                loss = loss_fct(logits.view(-1, self.num_labels).type(torch.LongTensor),
+                                     labels.view(-1).type(torch.LongTensor))
             outputs = (loss,) + outputs
 
         return outputs  # (loss), logits, (hidden_states), (attentions)
